backend/requirements.txt
fastapi
uvicorn
python-dotenv
gunicorn
hmac
hashlib
